Added report functions file and modified main code
