+++ main_gui.py
@@ -8,6 +8,9 @@ import os
 import pyautogui as pygui
 import customtkinter as ctk
 
+# Dependent files
+import report_functions as rf
+
 ## Main Window ##
 root = ctk.CTk()
 root.geometry("560x200")
@@ -50,6 +53,15 @@ def publish():
             locate_click(images_folder + "\\GW_zoomout.png")
             time.sleep(0.5)
     full(False,draw_report)
+
+def full(replace,draw):
+    for i in range(3):
+        rf.data_report(i,replace)
+        time.sleep(0.2)
+    rf.summary(replace)
+    time.sleep(0.2)
+    if draw == 1:
+        rf.drawing(replace)
 
 def open_outputs():
     # opens the .Outputs folder
@@ -238,114 +250,5 @@ project_button_show_1.grid(row=4,column=2,padx=(0,0),pady=(10,0),sticky='sw')
 project_label_2.grid(row=5,column=0,padx=(10,0),pady=(10,0),sticky='sw')
 project_button_2.grid(row=5,column=1,padx=(20,0),pady=(10,0),sticky='sw')
 project_button_show_2.grid(row=5,column=2,padx=(0,0),pady=(10,0),sticky='sw')
-
-## Functions ##
-
-# Background Tracing Based on CAD (legacy)
-    # was placed here, RIP
-
-def progress(dur):
-    # moves cursor from left to right end of screen to show wait time (legacy)
-    pygui.moveTo(0,45,duration=0)
-    pygui.moveTo(1915,45,duration=dur)
-
-def capture_change(x,y):
-    # checks if the cololur of a pixel at x,y has changed, then returns
-    check = True
-    while check:
-        im1 = pygui.screenshot()
-        im1_pixel = im1.getpixel((x,y))
-        time.sleep(1)
-        im2 = pygui.screenshot()
-        im2_pixel = im2.getpixel((x,y))
-        if not(im1_pixel == im2_pixel):
-            check = False
-    return
-
-def locate_click(image):
-    ## locate and click based on given image, in while loop
-    check = True
-    while check:
-        print_location = pygui.locateCenterOnScreen(image) # delay 1-2 sec
-        if not(print_location == None):
-            check = False
-    pygui.click(print_location[0],print_location[1])
-
-def data_report(ver,replace):
-    ## prints pipe/node/customer data report
-    pygui.moveTo(154,50,duration=0.2)   # to GW window
-    pygui.click()
-    pygui.hotkey("alt","r")             # Report tab
-    pygui.press(types[ver][0])
-    time.sleep(0.2)
-    locate_click(images_folder + "\\GW_print_button.png")
-    check = True
-    while check:
-        im = pygui.screenshot()
-        im_pixel = im.getpixel((screen_cx,screen_cy))
-        if im_pixel == (240, 240, 240):
-            check = False
-    time.sleep(0.2)
-    pygui.hotkey("alt","p") # prints report
-    capture_change(screen_cx,screen_cy)
-    pygui.typewrite(outputs_folder + "\\" + code + " - " + rev + " - " + types[ver] + " Data Report.pdf") # inputs name for report
-    pygui.hotkey("alt","s") # saves report
-    if replace == True:
-        pygui.press("y")
-    time.sleep(0.5)
-    pygui.press("y")
-    capture_change(screen_cx,screen_cy) # checks centre pixel change
-    pygui.hotkey("alt","c") # closes report window
-
-def summary(replace):
-    ## prints system summary report
-    pygui.moveTo(154,50,duration=0.2)   # to GW window
-    pygui.click()
-    pygui.hotkey("alt","r")             # Report tab
-    pygui.press("m")
-    pygui.hotkey("alt","c")
-    time.sleep(0.2)
-    locate_click(images_folder + "\\GW_print_button_2.png")
-    capture_change(screen_cx,screen_cy)
-    pygui.hotkey("alt","p") # prints report
-    capture_change(screen_cx,screen_cy)
-    pygui.typewrite(outputs_folder + "\\" + code + " - " + rev + " - System Summary Report.pdf")
-    pygui.hotkey("alt","s") # saves report
-    if replace == True:
-        pygui.press("tab")
-        pygui.press("enter")
-    time.sleep(1)             # waiting time for PDF to be fully published
-    pygui.hotkey("alt","c") # closes report window
-
-def drawing(replace):
-    ## prints noded drawing
-    locate_click(images_folder + "\\GW_button_side.png")
-    locate_click(images_folder + "\\GW_print_button.png")
-    pygui.press("enter")
-    time.sleep(0.5)
-    # note: saved view/window should be preselected at this stage
-    pygui.hotkey("alt","t") # navigates to Style pane
-    pygui.hotkey("alt","p") # plots drawing
-    capture_change(screen_cx,screen_cy)
-    # note: title block values should be filled in at this stage
-    pygui.hotkey("alt","c") # presses first "Continue"
-    time.sleep(0.3)
-    capture_change(screen_cx,screen_cy)
-    #progress(2)
-    pygui.hotkey("alt","c") # presses second "Continue"
-    pygui.typewrite(outputs_folder + "\\" + code + " - " + rev + " - Noded Drawing.pdf")
-    pygui.hotkey("alt","s") # saves report
-    if replace == True:
-        pygui.press("tab")
-        pygui.press("enter")
-
-def full(replace,draw):
-    for i in range(3):
-        data_report(i,replace)
-        time.sleep(0.2)
-    summary(replace)
-    time.sleep(0.2)
-    if draw == 1:
-        drawing(replace)
 
 root.mainloop()
